+++ src/openlmis-gs1/openlmis-scan-input.component.js
@@ -49,7 +49,7 @@
             bindings: {
                 mode: '@',
                 context: '<?',
-                onScan: '&'
+                onScan: '&?'
             }
         });